requirements.txt
# API access

requests
biopython # Entrez (PUBMED)
pandas
bs4
lxml

# Caching

diskcache
brotlicffi

# CLI logging
